% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{away.df}
\alias{away.df}
\title{Iowa State Men's Basketball data}
\format{
\subsection{\code{away.df}}{

A data frame with 13,126 rows and 4 columns:
\describe{
\item{game_id}{unique game identifier}
\item{secs_remaining}{seconds remaining in the game}
\item{home_score}{home team score}
\item{away}{away team score}
}
}
}
\source{
ESPN.com
}
\usage{
away.df
}
\description{
A data frame containing play-by-play data from the 2022-23 season
and beginning of 2023-24 season
}
\keyword{datasets}
