% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{live}
\alias{live}
\title{live}
\usage{
live(ppm = 3.5, var = 5, time = 20, current, threshold, line = NULL)
}
\arguments{
\item{ppm}{Points per minute combined of the two teams playing. This defaults to the NCAA 2021-22 & 22-23 average.}

\item{var}{Variance of PPM of the two teams. This defaults to the NCAA 2021-22 & 22-23 average.}

\item{time}{Time remaining in the game. Defaults to half game remaining (20min).}

\item{current}{current points scored}

\item{threshold}{scoring threshold to meet}

\item{line}{Optional - betting line (expected points scored)}
}
\value{
A list with the elements (if provided 'current' and 'threshold')
\item{exp.pts}{the expected points from the game}
\item{prob}{the probability it exceeds the provided threshold}
\item{price}{the estimated betting line}
}
\description{
live
}
\examples{
live(4.35, 6.9, time = 20, current = 89, threshold = 168.5, line = 150.5)
}
