% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{train}
\usage{
train(home.df, away.df)
}
\arguments{
\item{home.df}{play-by-play data frame for the home team. Data must be formatted in the following way:
col 1 = unique game identifier
col 2 = seconds remaining in the game
col 3 = team 1 score
col 4 = team 2 score}

\item{away.df}{play-by-play data frame for the home team. Data must be formatted in the following way:
col 1 = unique game identifier
col 2 = seconds remaining in the game
col 3 = team 1 score
col 4 = team 2 score}
}
\value{
PPM and variance of PPM to input into the total function
}
\description{
train
}
\examples{
train(home.df, away.df)
}
