% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotal.R
\name{plotTotal}
\alias{plotTotal}
\title{plotTotal}
\usage{
plotTotal(ppm = 3.5, var = 5, threshold, line)
}
\arguments{
\item{ppm}{points per minute combined of the two teams}

\item{var}{ppm variance of the two teams}

\item{threshold}{scoring threshold to meet}

\item{line}{betting line (expected points scored)}
}
\value{
prints out a hypothetical scoring flow chart
}
\description{
Plots a hypothetical game flow chart for points scored in a D1 Men's College Basketball game.
}
\examples{
plotTotal(4.96, 14.38, 168.5, 150.5)
}
