% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total}
\alias{total}
\title{total}
\usage{
total(ppm = 3.5, var = 5, line = NULL)
}
\arguments{
\item{ppm}{Points per minute combined of the two teams playing. This defaults to the NCAA 2021-22 & 22-23 average.}

\item{var}{Variance of PPM of the two teams. This defaults to the NCAA 2021-22 & 22-23 average.}

\item{line}{Optional - betting line (expected points scored)}
}
\value{
expected points as a scalar
\item{exp.pts}{the expected points from the game}
}
\description{
Calculates the total points expected to be scored in a standard D1 Men's College Basketball game.
}
\examples{
total(4.96, 14.38)
total(4.35, 6.9, line = 150.5)
}
